9C-1141-9 API: Recommendations and Widgets
We integrate the routes for "/recommendations" and "/widgets" in the
Applications API.
