+++ modules/api/src/main/scala/cards/nine/api/JsonFormats.scala
@@ -1,7 +1,6 @@
 package cards.nine.api
 
-import cards.nine.api.messages.GooglePlayMessages._
-import cards.nine.domain.application.{ Package, Widget }
+import cards.nine.domain.application.Package
 import spray.httpx.SprayJsonSupport
 import spray.json._
 
@@ -13,22 +12,6 @@ trait JsonFormats
     def read(json: JsValue): Package = Package(StringJsonFormat.read(json))
     def write(pack: Package): JsValue = StringJsonFormat.write(pack.value)
   }
-
-  implicit val apiGetRecommendationsByCategoryRequestFormat = jsonFormat2(ApiGetRecommendationsByCategoryRequest)
-
-  implicit val apiGetRecommendationsForAppsRequestFormat = jsonFormat4(ApiGetRecommendationsForAppsRequest)
-
-  implicit val googlePlayRecommendationFormat = jsonFormat7(ApiRecommendation)
-
-  implicit val apiGetRecommendationsResponseFormat = jsonFormat1(ApiGetRecommendationsResponse)
-
-  implicit val apiRankAppsByMomentsRequestFormat = jsonFormat4(ApiRankByMomentsRequest)
-
-  implicit val widgetFormat = jsonFormat2(Widget.apply)
-
-  implicit val apiRankedWidgetsByMomentFormat = jsonFormat2(ApiRankedWidgetsByMoment)
-
-  implicit val apiRankWidgetsResponseFormat = jsonFormat1(ApiRankWidgetsResponse)
 
 }
