+++ modules/api/src/main/scala/cards/nine/api/NineCardsApi.scala
@@ -1,19 +1,11 @@
 package cards.nine.api
 
 import akka.actor.ActorRefFactory
-import cards.nine.api.NineCardsDirectives._
-import cards.nine.api.NineCardsHeaders.Domain._
 import cards.nine.api.applications.ApplicationsApi
 import cards.nine.api.collections.CollectionsApi
 import cards.nine.api.rankings.RankingsApi
-import cards.nine.api.converters.Converters._
-import cards.nine.api.messages.GooglePlayMessages._
 import cards.nine.api.accounts.AccountsApi
-import cards.nine.api.utils.SprayMarshallers._
-import cards.nine.api.utils.SprayMatchers._
-import cards.nine.commons.NineCardsService.{ NineCardsService, Result }
 import cards.nine.commons.config.Domain.NineCardsConfiguration
-import cards.nine.domain.application.{ Category, PriceFilter }
 import cards.nine.processes._
 import cards.nine.processes.account.AccountProcesses
 import cards.nine.processes.applications.ApplicationProcesses
@@ -33,46 +25,13 @@ class NineCardsRoutes(
 ) {
 
   import Directives._
-  import JsonFormats._
 
-  lazy val nineCardsRoutes: Route = {
-    val accountsRoute = new AccountsApi().route
-    val applicationRoute = new ApplicationsApi().route
-    val collectionsRoute = new CollectionsApi().route
-    val rankingsRoute = new RankingsApi().route
-
+  lazy val nineCardsRoutes: Route =
     pathPrefix("apiDocs")(swaggerRoute) ~
-      accountsRoute ~
-      applicationRoute ~
-      collectionsRoute ~
-      rankingsRoute ~
-      pathPrefix("recommendations")(recommendationsRoute) ~
-      pathPrefix("widgets")(widgetRoute)
-  }
-
-  private[this] lazy val recommendationsRoute: Route =
-    nineCardsDirectives.authenticateUser { userContext ⇒
-      pathEndOrSingleSlash {
-        post {
-          entity(as[ApiGetRecommendationsForAppsRequest]) { request ⇒
-            nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
-              complete(getRecommendationsForApps(request, googlePlayContext, userContext))
-            }
-          }
-        }
-      } ~
-        pathPrefix(CategorySegment) { category ⇒
-          nineCardsDirectives.priceFilterPath { priceFilter ⇒
-            post {
-              entity(as[ApiGetRecommendationsByCategoryRequest]) { request ⇒
-                nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
-                  complete(getRecommendationsByCategory(request, category, priceFilter, googlePlayContext, userContext))
-                }
-              }
-            }
-          }
-        }
-    }
+      (new AccountsApi().route) ~
+      (new ApplicationsApi().route) ~
+      (new CollectionsApi().route) ~
+      (new RankingsApi().route)
 
   private[this] lazy val swaggerRoute: Route =
     // This path prefix grants access to the Swagger documentation.
@@ -82,57 +41,5 @@ class NineCardsRoutes(
     } ~ {
       getFromResourceDirectory("apiDocs")
     }
-
-  private[this] lazy val widgetRoute: Route =
-    nineCardsDirectives.authenticateUser { userContext ⇒
-      path("rank") {
-        post {
-          entity(as[ApiRankByMomentsRequest]) { request ⇒
-            complete(rankWidgets(request, userContext))
-          }
-        }
-      }
-    }
-
-  private type NineCardsServed[A] = cats.free.Free[NineCardsServices, A]
-
-  private[this] def getRecommendationsByCategory(
-    request: ApiGetRecommendationsByCategoryRequest,
-    category: Category,
-    priceFilter: PriceFilter,
-    googlePlayContext: GooglePlayContext,
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiGetRecommendationsResponse] =
-    applicationProcesses
-      .getRecommendationsByCategory(
-        category.entryName,
-        priceFilter,
-        request.excludePackages,
-        request.limit,
-        toMarketAuth(googlePlayContext, userContext)
-      )
-      .map(toApiGetRecommendationsResponse)
-
-  private[this] def getRecommendationsForApps(
-    request: ApiGetRecommendationsForAppsRequest,
-    googlePlayContext: GooglePlayContext,
-    userContext: UserContext
-  ): NineCardsService[NineCardsServices, ApiGetRecommendationsResponse] =
-    applicationProcesses
-      .getRecommendationsForApps(
-        request.packages,
-        request.excludePackages,
-        request.limitPerApp.getOrElse(Int.MaxValue),
-        request.limit,
-        toMarketAuth(googlePlayContext, userContext)
-      )
-      .map(toApiGetRecommendationsResponse)
-
-  private[this] def rankWidgets(
-    request: ApiRankByMomentsRequest,
-    userContext: UserContext
-  ): NineCardsServed[Result[ApiRankWidgetsResponse]] =
-    rankingProcesses.getRankedWidgets(request.location, request.items, request.moments, request.limit)
-      .map(toApiRankWidgetsResponse)
 
 }
