+++ modules/api/src/main/scala/cards/nine/api/applications/Api.scala
@@ -3,17 +3,17 @@ package cards.nine.api.applications
 import cards.nine.api.NineCardsDirectives._
 import cards.nine.api.NineCardsHeaders.Domain._
 import cards.nine.api.converters.Converters._
-import cards.nine.api.messages.GooglePlayMessages._
 import cards.nine.api.utils.SprayMarshallers._
 import cards.nine.api.utils.SprayMatchers._
 import cards.nine.commons.NineCardsService.{ NineCardsService, Result }
 import cards.nine.commons.config.Domain.NineCardsConfiguration
-import cards.nine.domain.application.{ BasicCard, FullCard, Package }
+import cards.nine.domain.application.{ BasicCard, Category, FullCard, Package, PriceFilter }
 import cards.nine.processes._
 import cards.nine.processes.account.AccountProcesses
 import cards.nine.processes.applications.ApplicationProcesses
 import cards.nine.processes.rankings.RankingProcesses
 import cards.nine.processes.NineCardsServices._
+import cats.free.Free
 import scala.concurrent.ExecutionContext
 import spray.routing._
 
@@ -31,7 +31,10 @@ class ApplicationsApi(
   import JsonFormats._
   import messages._
 
-  lazy val route: Route = pathPrefix("applications")(details ~ categorize ~ rank ~ search)
+  lazy val route: Route =
+    pathPrefix("applications")(details ~ categorize ~ rank ~ search) ~
+      pathPrefix("recommendations")(recommendationsRoute) ~
+      pathPrefix("widgets")(widgetRoute)
 
   private[this] val details: Route =
     pathPrefix("details") {
@@ -108,6 +111,41 @@ class ApplicationsApi(
       }
     }
 
+  private[this] lazy val recommendationsRoute: Route =
+    nineCardsDirectives.authenticateUser { userContext ⇒
+      pathEndOrSingleSlash {
+        post {
+          entity(as[ApiGetRecommendationsForAppsRequest]) { request ⇒
+            nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
+              complete(getRecommendationsForApps(request, googlePlayContext, userContext))
+            }
+          }
+        }
+      } ~
+        pathPrefix(CategorySegment) { category ⇒
+          nineCardsDirectives.priceFilterPath { priceFilter ⇒
+            post {
+              entity(as[ApiGetRecommendationsByCategoryRequest]) { request ⇒
+                nineCardsDirectives.googlePlayInfo { googlePlayContext ⇒
+                  complete(getRecommendationsByCategory(request, category, priceFilter, googlePlayContext, userContext))
+                }
+              }
+            }
+          }
+        }
+    }
+
+  private[this] lazy val widgetRoute: Route =
+    nineCardsDirectives.authenticateUser { userContext ⇒
+      path("rank") {
+        post {
+          entity(as[ApiRankByMomentsRequest]) { request ⇒
+            complete(rankWidgets(request, userContext))
+          }
+        }
+      }
+    }
+
   private[this] def getAppsInfo[T](
     request: ApiAppsInfoRequest,
     googlePlayContext: GooglePlayContext,
@@ -161,5 +199,44 @@ class ApplicationsApi(
   ): cats.free.Free[NineCardsServices, Result[ApiRankAppsResponse]] =
     rankingProcesses.getRankedAppsByMoment(request.location, request.items, request.moments, request.limit)
       .map(toApiRankAppsResponse)
+
+  private[this] def getRecommendationsByCategory(
+    request: ApiGetRecommendationsByCategoryRequest,
+    category: Category,
+    priceFilter: PriceFilter,
+    googlePlayContext: GooglePlayContext,
+    userContext: UserContext
+  ): NineCardsService[NineCardsServices, ApiGetRecommendationsResponse] =
+    applicationProcesses
+      .getRecommendationsByCategory(
+        category.entryName,
+        priceFilter,
+        request.excludePackages,
+        request.limit,
+        toMarketAuth(googlePlayContext, userContext)
+      )
+      .map(toApiGetRecommendationsResponse)
+
+  private[this] def getRecommendationsForApps(
+    request: ApiGetRecommendationsForAppsRequest,
+    googlePlayContext: GooglePlayContext,
+    userContext: UserContext
+  ): NineCardsService[NineCardsServices, ApiGetRecommendationsResponse] =
+    applicationProcesses
+      .getRecommendationsForApps(
+        request.packages,
+        request.excludePackages,
+        request.limitPerApp.getOrElse(Int.MaxValue),
+        request.limit,
+        toMarketAuth(googlePlayContext, userContext)
+      )
+      .map(toApiGetRecommendationsResponse)
+
+  private[this] def rankWidgets(
+    request: ApiRankByMomentsRequest,
+    userContext: UserContext
+  ): Free[NineCardsServices, Result[ApiRankWidgetsResponse]] =
+    rankingProcesses.getRankedWidgets(request.location, request.items, request.moments, request.limit)
+      .map(toApiRankWidgetsResponse)
 
 }
